organize code structure
